requirements.txt
aniso8601==9.0.1
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
